LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-type-to-search.tentative-expected.txt

alpha bravo charlie

FAIL Type to search should focus but not select an option until selection is confirmed. assert_equals: The `bravo` option should be focused after pressing `b`. expected Element node <option id="bravo">bravo</option> but got Element node <select id="target">
  <option></option>
  <option>alpha<...
